feat: add new command to instantly creation with params
