+++ package.json
@@ -34,7 +34,7 @@
     "test": "node --test",
     "test:watch": "node --test --watch",
     "test:coverage": "node --test --experimental-test-coverage --test-reporter=lcov --test-reporter-destination=lcov.info",
-    "dev": "yarn build && clear && node src/main.js gen",
+    "dev": "yarn build && clear && node src/main.js",
     "dev:npx": "yarn build && npm link && clear && npx clingon gen"
   },
   "dependencies": {
